Requires scripts re-think vecdraw disposal.
Vecdraws used to get disposed immediately upon
setting the variable to something else, or
that variable going out of scope, but that had
lots of bugs.  But in the simplest scenario it
worked:

```
set v1 to vecdraw(v(0,0,0),v(10,0,0),red,"arrow",1,true,1).
wait 1.
set v1 to 0. // vecdraw immediately goes away.
```

But it failed in cases where the vecdraw was not stored
in a simple identifier but in a collection:

```
set L to list().
set v1 to vecdraw(v(0,0,0),v(10,0,0),red,"arrow",1,true,1).
L:ADD( v1 ).
set v1 to 0. // vecdraw goes away even though it isn't really orphaned.
             // because it's still in the L list.
```

This was essentially from us trying to perform a crude
homebrewed garbage collection that didn't really work well.
This change in this commit causes us to fall back on relying
on the underlying GC in the system itself instead.  The
advantage is that it's always done correctly.  The disadvantage
is that relying on the system's GC means the arrows don't always
disappear immediately, and tend to only go away when other memory
gets allocated elsewhere, thus kicking the GC into waking up and
doing something.  After this change, scripts should just be in the
habit of always using the ``:show`` suffix to turn off a vecdraw
they're about to get rid of, or calling ``clearvecdraws()`` to
force them to get removed.